requirements.txt
langchain
openai
umap-learn 
scikit-learn 
langchain_community 
tiktoken 
langchain-openai 
langchainhub 
chromadb 
langchain-anthropic
streamlit
numpy
pandas
langchainhub
langgraph 
tavily-python 
langchain-mistralai 
gpt4all 
llama-cpp-python
python-dotenv
google-generativeai
langchain_google_genai
bs4
matplotlib
tqdm
deeplake[enterprise]
#!CMAKE_ARGS="-DLLAMA_METAL=on" FORCE_CMAKE=1 pip install -U llama-cpp-python --no-cache-dir